+++ packages/terra-button/src/terra-dev-site/doc/button/Button.doc.jsx
@@ -65,6 +65,7 @@ const DocPage = () => (
         componentSrc: ButtonSrc,
       },
     ]}
+    customPropsTable={<div>Custom props table</div>}
   />
 );