+++ packages/terra-doc-template/src/terra-dev-site/test/doc-template/DefaultDocTemplate.test.jsx
@@ -3,6 +3,7 @@ import DocTemplate from '../../../DocTemplate';
 import readme from './common/TestReadme.md';
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions, import/no-duplicates */
 import TestComponentExample from './common/TestComponentExample';
+import CustomPropsTable from './common/CustomPropsTable';
 import exampleSrc from '!raw-loader!../../../../src/terra-dev-site/test/doc-template/common/TestComponentExample';
 import testComponentSrc from '!raw-loader!../../../../src/terra-dev-site/test/doc-template/common/TestComponent';
 /* eslint-enisable import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
@@ -30,6 +31,7 @@ const Index = () => {
         readme={readme}
         srcPath="https://github.com/cerner/terra-core/tree/master/packages/terra-doc-template"
         propsTables={propsTables}
+        customPropsTable={<CustomPropsTable />}
         examples={examples}
       />
     </div>